task 0

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t n = 0;
+	unsigned int n = 0;
 	const list_t *temp = h;
 
 	while (temp)
